put canvas view stuff in its own file
